Added profile and minor changes to the frontend

+++ backend/src/modules/tasks/task.controller.js
@@ -2,8 +2,9 @@ const taskService = require('./task.service.js');
 
 exports.createTask = async (req, res, next) => {
     try {
+        const user_id = req.user.user_id;
         const { task_title, task_description, task_status, category, priority } = req.body;
-        const task = await taskService.createTask({task_title, task_description, task_status, category, priority});
+        const task = await taskService.createTask({user_id, task_title, task_description, task_status, category, priority});
 
         res.status(201).json(task);
     } catch (err) {
@@ -11,9 +12,10 @@ exports.createTask = async (req, res, next) => {
     } 
 };
 
-exports.getAllTasks = async (req, res, next) => {
+exports.getTasksByUser = async (req, res, next) => {
     try {
-        const tasks = await taskService.getAllTasks();
+        const user_id = req.user.user_id;
+        const tasks = await taskService.getTasksByUser({user_id});
 
         res.status(200).json(tasks);
     } catch (err) {
@@ -36,7 +38,8 @@ exports.updateTask = async (req, res, next) => {
     try {
         const { new_task_title, new_task_description, new_task_status, new_category, new_priority } = req.body;
         const { task_id } = req.params;
-        const task = await taskService.updateTask({task_id, new_task_title, new_task_description, new_task_status, new_category, new_priority});
+        const user_id = req.user.user_id;
+        const task = await taskService.updateTask({user_id, task_id, new_task_title, new_task_description, new_task_status, new_category, new_priority});
 
         res.status(200).json(task);
     } catch (err) {
@@ -46,8 +49,9 @@ exports.updateTask = async (req, res, next) => {
 
 exports.deleteTask = async (req, res, next) => {
     try {
+        const user_id = req.user.user_id;
         const { task_id } = req.params;
-        await taskService.deleteTask({task_id});
+        await taskService.deleteTask({user_id, task_id});
 
         res.status(200).json({ message: 'Задача успешно удалена' });
     } catch (err) {
